Create Exercise5_c++
I learned some basic file operations: how to create a .txt file, open and close it, and how to save its content into variables. I learned comparing strings was easier than I thought in C++. 

The prompt said to break in if the password is wrong and I couldn't think of an advanced way to "hack in" if you enter the wrong input so I just made the program display the file contents with the correct answers if they guess the input incorrectly.