Update Exercise6_c++
Lots of trouble... first of all I was making my for loops too complicated by trying to add in the 3.0 in the loop instead of just computer the decimal place. I still struggle predicting what loops due and feel like I just throw random changes in their hoping I'll get the right answer. But I am not sure if C++ can compute a decimal place precise enough so it will match, if the series I chose just can't get accurate, or if there is something off in the code!

Hopefully I will get a better grasp on variables for big and precise numbers someday and will come back to this.