Create Exercise10_c++(validate phone #s)
From my research for this one validating input is tricky in C++. There is the regex library which is over my head but I hope to come back to this. I could not figure out how to expand upon this and was just copy pasting so hoping to return to this problem.