Create Exercise15_c++(right justify file)
I wanted to do this to modify an existing file and add the right justification to it but it turns out it's harder than it seems and kept clearing the file contents. You can append a file with instructions but not begin at the start and keep the contents. It was suggested to copy the contents of this file to another then back again, which were only confusing examples with sometimes outdated ".h" headers in them.