Add files via upload
Tried to make the splitting of the words wrap. Could figure it out without adding the stars on the sides using rfind, but to add the stars I was trying to make new words of a certain length and add them to an array of string pointers... couldn't figure out how to make sure it split at a space, under the length limit, and then save that as a string - trouble figuring out how to word what needs to be done.